[backport] switch before rope refactor
