+++ vllm/model_executor/models/ernie45_vl_moe.py
@@ -41,7 +41,7 @@ from vllm.model_executor.layers.linear import (QKVParallelLinear,
                                                RowParallelLinear)
 from vllm.model_executor.layers.logits_processor import LogitsProcessor
 from vllm.model_executor.layers.quantization import QuantizationConfig
-from vllm.model_executor.layers.rotary_embedding.ernie45_vl_rope import (
+from vllm.model_executor.layers.rotary_embedding import (
     Ernie4_5_VLRotaryEmbedding)
 from vllm.model_executor.layers.vocab_parallel_embedding import (
     ParallelLMHead, VocabParallelEmbedding)
